I optimized the calculateDocFreqs function with this outcome:
total time (function time)
40s (112 ms) - fast current algorithm
126s (84,681 ms) - slow slightly optimized version that we created during our meeting.
167s (120,442 ms) - slowest original
@clohertyg  I've renamed the old original function calculateDocFreqsSlow() so that it isn't used but you can compare it to the new code and see the changes.